Move tree choice into tile gen, not calc on display
