+++ src/game/overview/Display.js
@@ -1,70 +1,4 @@
 import ROT from "rot-js";
-const trees = [
-  {
-    "symbol": "↟",
-    "prob": 20
-  },
-  // {
-  //   "symbol": "🌲",
-  //   "prob": 20
-  // },
-  // {
-  //   "symbol": "🌳",
-  //   "prob": 15
-  // },
-  {
-    "symbol": "⇞",
-    "prob": 12
-  },
-  {
-    "symbol": "↑",
-    "prob": 8
-  },
-  {
-    "symbol": "⇈",
-    "prob": 6
-  },
-  {
-    "symbol": "⭫",
-    "prob": 5
-  },
-  {
-    "symbol": "𐇲",
-    "prob": 5
-  },
-  {
-    "symbol": "⇡",
-    "prob": 4
-  },
-  {
-    "symbol": "⇑",
-    "prob": 4
-  },
-  {
-    "symbol": "⥉",
-    "prob": 2
-  },
-  {
-    "symbol": "⭎",
-    "prob": 2
-  },
-  {
-    "symbol": "⤒",
-    "prob": 2
-  },
-  {
-    "symbol": "⭂",
-    "prob": 1
-  },
-  {
-    "symbol": "⭜",
-    "prob": 1
-  }
-]
-
-trees.sort((a, b) => b.prob - a.prob);
-let total = trees.reduce((acc, el, i) => (el.prob += acc), 0);
-
 ROT.Display.prototype.drawWorld = function(world, playerPos) {
   let curOpts = display.getOptions();
   let top = Math.floor(playerPos.y - curOpts.height / 2);
@@ -81,12 +15,7 @@ ROT.Display.prototype.drawWorld = function(world, playerPos) {
       } else if (tile.creatures.length) {
         display.draw(j, i, "@");
       } else {
-        let rand = ~~(Math.random() * total);
-        let tree = 0;
-        while (rand >= trees[tree + 1].prob) {
-          tree++
-        }
-        display.draw(j, i, trees[tree].symbol, "#086623");
+        display.draw(j, i, tile.display, "#086623");
       }
       j++;
     }
